+++ odoo/local-src/connector_qoqa/exception.py
@@ -14,16 +14,14 @@ class QoQaAPIError(QoQaError):
     """ Error with the QoQa API. """
 
 
-class QoQaAPIAuthError(QoQaAPIError):
-    """ Authentication error with the QoQa API. """
-
-
 class QoQaResponseNotParsable(QoQaAPIError):
     """ Happens when a response from the QoQa API is not parsable. """
 
 
 class QoQaResponseError(QoQaAPIError):
     """ The API responded but gave details about errors """
+
+    no_reason = u'Unknown error'
 
     def __init__(self, errors):
         """
@@ -35,12 +33,19 @@ class QoQaResponseError(QoQaAPIError):
 
     def __str__(self):
         if not self.errors:
-            return u'Unknow error'
+            return self.no_reason
         else:
             return ','.join([
                 u"[code {0}] {1}: {2}".format(code, title, detail)
                 for code, title, detail in self.errors
             ])
+
+
+class QoQaAPIAuthError(QoQaResponseError):
+    """ Authentication error with the QoQa API. """
+
+    no_reason = (u'The call could not be completed because it was not '
+                 u'authenticated or the authentication rejected it.')
 
 
 class OrderImportRuleRetry(RetryableJobError):