+++ HISTORY.rst
@@ -23,6 +23,7 @@ latest (unreleased)
 
 **Features and Improvements**
 
+* When the authentication to Q4 API fails, shows the error message returned by the API
 * Delete product and variants on QoQa API when product is disabled on Odoo (BSQOQ-6)
 
 **Bugfixes**